updated testbed interface
