+++ pyamg/testbed/__init__.py
@@ -1,1 +1,7 @@
+"""PyAMG Matrix Test Bed"""
 
+from . import matrices
+from .matrices import examples, get_mat
+
+# Include matrix testbed interface function
+__all__ = ['examples, get_mat']
